requirements.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==3.6.2
bcrypt==4.0.1
blinker==1.7.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
cryptography==41.0.5
databases==0.8.0
dnspython==2.4.2
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.95.0
Flask==2.3.3
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.3
mysqlclient==2.2.0
passlib==1.7.4
pipreqs==0.4.13
psycopg2==2.9.9
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.6.0
PyMySQL==1.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
radar==0.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
SQLAlchemy-Utils==0.41.1
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.7
uvicorn==0.21.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
win32-setctime==1.1.0
wincertstore==0.2
yarg==0.1.9
